+++ pages/index.tsx
@@ -1,14 +1,14 @@
 import Layout from '@/components/common/Layout';
-import ExchangeRateBox from '@/components/ExchangeRate';
+import ExchangeRateBox from '@/components/dividend/ExchangeRate';
 import ModeController from '@/components/ModeController';
 import CommonCenter from '@/components/common/Center';
 import CommonHeader from '@/components/common/Header';
-import DetailStocksSection from '@/components/PortFolioSections/DetailStocksSection';
-import SimpleDividentScheduleSection from '@/components/PortFolioSections/SimpleDividentScheduleSection';
-import ChartSection from '@/components/PortFolioSections/ChartSection';
-import TotalStatisticsSection from '@/components/PortFolioSections/TotalStatisticsSection';
+import DetailStocksSection from '@/components/portfolio/portfolioSections/DetailStocksSection';
+import SimpleDividentScheduleSection from '@/components/portfolio/portfolioSections/SimpleDividentScheduleSection';
+import ChartSection from '@/components/portfolio/portfolioSections/ChartSection';
+import TotalStatisticsSection from '@/components/portfolio/portfolioSections/TotalStatisticsSection';
 import { useMyPortFolio } from '@/hook/useMyPortFolio';
-import NotifyMessageSection from '@/components/PortFolioSections/NotifyMessageSection';
+import NotifyMessageSection from '@/components/portfolio/portfolioSections/NotifyMessageSection';
 import { CircularProgress } from '@mui/material';
 
 import Head from 'next/head';
